requirements.txt
asgi-redis==1.0.0
asgiref==2.0.1
asn1crypto==0.24.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
braintree==3.43.0
certifi==2018.1.18
cffi==1.11.5
channels==1.0.3
channels-api==0.4.1
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.2.2
daphne==1.0.3
Django==2.0.1
djangorestframework==3.7.7
drfdocs==0.0.11
functions==0.7.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
more-itertools==4.1.0
msgpack==0.5.6
msgpack-python==0.4.8
pluggy==0.6.0
py==1.5.3
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.6.1
pypiwin32==219
pytest==3.5.0
pytz==2018.3
redis==2.10.5
requests==2.18.4
schemas==0.7.1
six==1.11.0
stripe==1.79.1
suds-jurko==0.6
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
websocket-client==0.47.0
xmltodict==0.11.0
zope.interface==4.4.3
